feat: Add option to choose a size / resolution for slides (#16)
There is a problem that the slides don't scale and they can't exactly tell what it is when looking at the big window. This PR makes it possible to set a proper resolution which then scales up or down on changing the window size by keeping it constant on any dimensioned device.